Create TL-GinjoSake.xml
Shortened the Ginjo barrel's description. "Sake" is common knowledge in the Western World, and is recognized as a rice-based alcohol.
Added punctuation to bill descriptions and job strings to maintain consistency with later entries.
Added space between item and quantity.
Changed job string label for Ginjo Sake in order to reflect it coming out of a barrel.