Create TL-Recipes_Cookingtable.xml
Translated and localized multiple fragments of Japanese text.
Added spacing between item names and quantities. This helps the player see the item and amount they are making.
Changed "Seasoned meat" label to "Season meat." This adds variety to the bill last, and accurately describe the bill itself.
Changed "Frozen Mochi" label to "Freezer Mochi." Since RimWorld recognizes items as "frozen" when they are below a certain temperature, I thought it prudent to change this label. Freezer Mochi is not always frozen. Rather, it is Mochi that has been prepared to be frozen.