Create TL-Recipes_Food_Drink.xml
Changed tense of "dissolve" and "add." This shifts the sentence's meaning from describing what is happening to what can happen.
Added comedic text to Soy milkshake's description. Inserting such text helps the mod in several ways: it breaks the flow of mechanical descriptions; it humanizes the modders by showing they have a sense of humor; and it might get a chuckle from the player. Such asides-as an out of character joke or anecdote is called-should be kept to a minimum. Doing them to frequently destroys their novelty.