Create TL-Recipes_Food_MealFine.xml
Expanded abbreviations such as "BR, WR, and TKG," to fit Rimworld's base game aesthetic. TKG was specifically changed to allow players to learn the formal name of the dish. According to research TKG is a common acronym for the dish in Japan. I will add this cultural note to the item's proper description. I understand the desire to avoid visual clutter, but using the full words avoids any ambiguity that might confuse some players. Moreover, I kept the abbreviations in every jobstring, thus allowing the player to become familiar with those terms that way.
Added spacing between item names and quantities.
Added periods to jobstrings.
Researched and expanded the description of Sashimi appropriately.
Translated and localized the descriptions for Takikomi Gohan, Tendon, Okonomiyaki, Udon, and Oden.