Create TL-Recipes_Food_MealSimple.xml
Added spacing between item names and quantities.
Added a period to the end of each jobstring.
As with TKG, I removed acronyms to allow the player to see what they are cooking. Moreover, the base game omits the use of acronyms, which means there is no precedent for their inclusion.
Made adjustments to the descriptions of BR and WR Gohan based on research. Included Japanese terms, such as kome and asa, for those interested in Japanese language and culture.
Slightly adjusted Tsukudani's description.
Translated and localized the description for Korokke.