Create TL-Recipes_Food_Survival.xml
Added spacing between item names and quantities.
Translated and localized the descriptions and jobstrings for Alpha Rice, Hardtack, and Chimaki.
Changed the name of "α Rice" to Alpha Rice. The "α" symbol looked too similar to the letter "a" when in-game. This had the effect of making the player believe they were cooking "a Rice" instead of Alpha Rice. 
Made descriptions for Hardtack and Chimaki based on research.