Create TL-Recipes_RiceMillingMachine.xml
Added spacing between item names and quantities.
Adjusted jobstrings.
Translated and localized various entries that were left in Japanese.
Adjusted several labels to remove symbols from them. Symbols such as "⇒" can be difficult to see in-game. This change was made to prevent possible player confusion.
Corrected "Outher" to "Outer."
Though objects such as "White rice," "Outer hull," and "Rice bran" are mundane enough to not be capitalized, I chose to capitalize the first word of each object in order to increase their visibility within a sentence. Players often scan descriptions for keywords, so it helps to make sure the ingredients and results are both highlighted in some way.